requirements.txt
streamlit
google-generativeai
youtube-transcript-api
fpdf
python-dotenv